# pad counters