--- a/BitFaster.Caching/Padding.cs
+++ b/BitFaster.Caching/Padding.cs
@@ -4,7 +4,6 @@ using System.Text;
 
 namespace BitFaster.Caching
 {
-    // https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/Internal/Padding.cs
     internal class Padding
     {
 #if TARGET_ARM64 || TARGET_LOONGARCH64